add USGS file and sensor interface
