+++ stormevents/usgs/__init__.py
@@ -1,6 +1,6 @@
-from stormevents.usgs.highwatermarks import (
-    FloodEvent,
-    StormFloodEvent,
+from stormevents.usgs.events import (
+    USGS_Event,
     usgs_flood_events,
     usgs_flood_storms,
+    USGS_StormEvent,
 )
